requirements.txt
fastapi>=0.109.0,<1.0
uvicorn[standard]>=0.27.0,<1.0
httpx>=0.26.0,<1.0
pyyaml>=6.0,<7.0
zeroconf>=0.131.0,<1.0
watchdog>=4.0.0,<7.0
